Refactor firebase feature
