+++ src/features/firebase/analytics/index.ts
@@ -1,0 +1,4 @@
+import Analytics from '@react-native-firebase/analytics';
+
+export const logEvent = (eventName: string, data: any) =>
+  Analytics().logEvent(eventName, data);
