+++ src/features/firebase/database/tools.ts
@@ -1,0 +1,6 @@
+// @ts-ignore
+import {DataSnapshot} from '@firebase/database';
+
+export const parseDatabaseSnapshot = (data?: DataSnapshot) => {
+  return data?.val();
+};
