+++ src/features/index.ts
@@ -1,6 +1,6 @@
 export {default as navigation} from './navigation';
 import * as system from './system';
 import * as store from './store';
-export * from './firebase';
+import * as firebase from './firebase';
 
-export {system, store};
+export {system, store, firebase};
